+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF5NV6EK91F86R33CP4PVZ1N"
+        "01KF9VM0EVQB8Y45CPRMEH2PYK"
     )
     r"""The unique identifier of the data source"""