+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF5NV6EVM04CZXM9R51BXWCC"
+        "01KF9VM0F088384XRMPBT4S1V7"
     )
     r"""The unique identifier of the data source"""