+++ packages/orq-rc/RELEASES.md
@@ -4749,3 +4749,13 @@ Based on:
 - [python v4.2.0-rc.37] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.37] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.37 - packages/orq-rc
+
+## 2026-01-19 00:51:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.38] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.38 - packages/orq-rc